add spotbugs-maven-plugin, fix some warnings

+++ web/src/main/java/com/developmentontheedge/be5/web/ParametersAccess.java
@@ -46,7 +46,7 @@ public interface ParametersAccess
     {
         String s = get(parameter);
 
-        return s != null ? Long.parseLong(s) : defaultValue;
+        return s != null ? (Long) Long.parseLong(s) : defaultValue;
     }
 
     /**
